Update policies and actions flow according to sleeping instances
